docs/tools/pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "cocotb-dv-guide-tooling"
version = "0.1.0"
description = "Internal authoring tools for the Cocotb design verification guide repository."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "graphviz>=0.20",
  "pydantic>=2.8",
  "PyYAML>=6.0.2",
  "rich>=14.0.0",
  "schemdraw>=0.19",
  "typer>=0.16.0",
  "wavedrom>=2.0.3.post3",
]

[project.optional-dependencies]
dev = [
  "Pygments>=2.19.0",
  "pytest>=8.3.0",
  "ruff>=0.11.0",
]

[project.scripts]
diagramgen = "cocotb_dv_guide.diagramgen.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cocotb_dv_guide"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]
